+++ Providers/Authentik/Private/Test-AuthentikPrerequisite.ps1
@@ -46,6 +46,7 @@ function Test-AuthentikPrerequisite {
         $dataPath = Get-AuthentikDataPath
         $requiredFiles = @(
             'AuthentikGroups.csv', 'AuthentikUsers.csv', 'AuthentikRoles.csv', 'AuthentikApplications.csv', 'AuthentikOutposts.csv',
+            'AuthentikStages.csv', 'AuthentikFlows.csv',
             'AuthentikScopeMappings.csv', 'AuthentikEntitlements.csv', 'AuthentikPolicies.csv',
             'AuthentikNotificationRules.csv', 'AuthentikBindings.csv', 'AuthentikTokens.csv',
             'AuthentikInvitations.csv'